# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "API documentation for Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Blog/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "My Comments"
                ],
                "summary": "My all Comments and all available Reactos",
                "description": "Get all comments of an authenticated user with:\n     *                  - the post of each comment\n     *                  - the post author\n     *                  - reactos of the comment\n     *                  - reactos of the post\n     *                  - and all available reactos",
                "operationId": "0e6682d8bd58b950fba21edbfa745800",
                "responses": {
                    "200": {
                        "description": "Paginated list of comments with comment, reactos and all reactos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All Comments User with the post of ever comment and reactos of every Comment Retrieved Successfully, and all available Reactos"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/comments?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/comments?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/comments?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/comments"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": "http://localhost/Blog/public/api/comments?page=1",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "1"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "reactos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Show my comments from deactivated account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Account is deactivated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "My Comments"
                ],
                "summary": "Create My Comment",
                "description": "Authenticated user Create Comment",
                "operationId": "044cb748b7d21a58b1cd59af39307eae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "lang",
                                    "post_id"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "anther post"
                                    },
                                    "lang": {
                                        "type": "string",
                                        "enum": [
                                            "ar",
                                            "en",
                                            "ur",
                                            "sp"
                                        ],
                                        "example": "ar"
                                    },
                                    "post_id": {
                                        "description": "exists('posts', 'id')",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "store comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Store comment from deactivated account or unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Account is deactivated or Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{comment}": {
            "put": {
                "tags": [
                    "My Comments"
                ],
                "summary": "Update My Comment",
                "description": "Authenticated user update Comment",
                "operationId": "ca2fa661e113bef3fdf348410b7c12fd",
                "parameters": [
                    {
                        "name": "comments",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "lang"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "anther post"
                                    },
                                    "lang": {
                                        "type": "string",
                                        "enum": [
                                            "ar",
                                            "en",
                                            "ur",
                                            "sp"
                                        ],
                                        "example": "ar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Update comment from deactivated account or unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Account is deactivated or Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "My Comments"
                ],
                "summary": "Delete My Comment",
                "description": "Authenticated user delete Comment",
                "operationId": "9c047c228e7e0fd8bc175b6be698bd5d",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CommentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lang": {
                        "type": "string",
                        "example": "user lang"
                    },
                    "body": {
                        "type": "string",
                        "example": "comment content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "post": {
                        "$ref": "#/components/schemas/PostResource"
                    },
                    "reactos": {
                        "description": "Loaded only if relation reactos is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReactResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "PostResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lang": {
                        "type": "string",
                        "example": "user lang"
                    },
                    "body": {
                        "type": "string",
                        "example": "post content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    },
                    "reactos": {
                        "description": "Loaded only if relation reactos is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReactResource"
                        }
                    }
                },
                "type": "object"
            },
            "ReactResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "tag name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "posts": {
                        "description": "Loaded only if relation posts is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostResource"
                        }
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}